Refactored to create tag layout mixin to simplify the two manipulations components
