+++ app/javascript/custom-tagged-plate/components/TagLayoutManipulations.vue
@@ -102,7 +102,6 @@
           ref="offsetTagsByComponent"
           :number-of-tags="numberOfTags"
           :number-of-target-wells="numberOfTargetWells"
-          :tags-per-well="tagsPerWell"
           @tagoffsetchanged="tagOffsetChanged"
         />
       </b-col>
@@ -126,9 +125,8 @@
 
 <script>
 import PlateScan from 'shared/components/PlateScan'
-import TagGroupsLookup from 'shared/components/TagGroupsLookup.vue'
-import TagOffset from './TagOffset.vue'
 import { checkState, checkQCableWalkingBy, aggregate } from 'shared/components/plateScanValidators'
+import TagLayout from 'custom-tagged-plate/components/mixins/TagLayout'
 
 /**
  * Allows the user to select tags and arrange their layout on the plate.
@@ -147,53 +145,14 @@ import { checkState, checkQCableWalkingBy, aggregate } from 'shared/components/p
 export default {
   name: 'TagLayoutManipulations',
   components: {
-    'lb-plate-scan': PlateScan,
-    'lb-tag-groups-lookup': TagGroupsLookup,
-    'lb-tag-offset': TagOffset
+    'lb-plate-scan': PlateScan
   },
-  props: {
-    // A devour API object. eg. new devourClient(apiOptions)
-    // Passed through to the tag groups lookup and tag plate scan child
-    // components.
-    api: {
-      type: Object,
-      required: true
-    },
-    // The current number of useable tags, calculated by the parent component
-    // and used to determine tag offset limits.
-    numberOfTags: {
-      type: Number,
-      default: 0
-    },
-    // The number of target wells, calculated by the parent component and
-    // used to determine the tag offset limits.
-    numberOfTargetWells: {
-      type: Number,
-      default: 0
-    },
-    // The tags per well number, determined by the plate purpose and used here
-    // to determine what tag layout walking by options are available.
-    tagsPerWell: {
-      type: Number,
-      default: 1
-    },
-  },
+  mixins: [TagLayout],
   data () {
     return {
-      tagGroupsList: {}, // holds the list of tag groups once retrieved
-      tagPlate: null, // holds the tag plate once scanned
       tagPlateWasScanned: false, // flag to indicate a tag plate was scanned
       tagPlateScanDisabled: false, // flag to indicate the plate scan was disabled
-      tag1GroupId: null, // holds the id of tag group 1 once selected
-      tag2GroupId: null, // holds the id of tag group 2 once selected
       walkingBy: 'manual by plate', // holds the chosen tag layout walking by option
-      direction: 'column', // holds the chosen tag layout direction option
-      offsetTagsBy: 0, // holds the entered tag offset number
-      nullTagGroup: { // null tag group object used in place of a selected tag group
-        uuid: null, // uuid of the tag group
-        name: 'No tag group selected', // name of the tag group
-        tags: [] // array of tags in the tag group
-      }
     }
   },
   computed: {
@@ -205,34 +164,8 @@ export default {
         { value: 'wells of plate', text: 'By Plate (Fixed)' }
       ]
     },
-    directionOptions() {
-      return [
-        { value: null, text: 'Please select a Direction Option...' },
-        { value: 'row', text: 'By Rows' },
-        { value: 'column', text: 'By Columns' },
-        { value: 'inverse row', text: 'By Inverse Rows' },
-        { value: 'inverse column', text: 'By Inverse Columns' }
-      ]
-    },
     tagGroupsDisabled() {
       return (typeof this.tagPlate != 'undefined' && this.tagPlate !== null)
-    },
-    tag1Group() {
-      return this.tagGroupsList[this.tag1GroupId] || this.nullTagGroup
-    },
-    tag2Group() {
-      return this.tagGroupsList[this.tag2GroupId] || this.nullTagGroup
-    },
-    coreTagGroupOptions() {
-      return Object.values(this.tagGroupsList).map(tagGroup => {
-        return { value: tagGroup.id, text: tagGroup.name }
-      })
-    },
-    tag1GroupOptions() {
-      return [{ value: null, text: 'Please select an i7 Tag 1 group...' }].concat(this.coreTagGroupOptions.slice())
-    },
-    tag2GroupOptions() {
-      return [{ value: null, text: 'Please select an i5 Tag 2 group...' }].concat(this.coreTagGroupOptions.slice())
     },
     scanValidation() {
       return () => {
@@ -241,15 +174,6 @@ export default {
     }
   },
   methods: {
-    tagGroupsLookupUpdated(data) {
-      if(!data || data.state === 'searching') {
-        return
-      } else if(data.state === 'valid') {
-        this.tagGroupsList = { ...data.tagGroupsList }
-      } else {
-        console.log('Tag Groups lookup error: ', data['state'])
-      }
-    },
     updateTagPlateScanDisabled() {
       if(this.tagPlateWasScanned) {
         this.tagPlateScanDisabled = false
@@ -298,22 +222,6 @@ export default {
     },
     tagGroupInput() {
       this.updateTagPlateScanDisabled()
-    },
-    tagOffsetChanged(tagOffset) {
-      this.offsetTagsBy = tagOffset
-      this.updateTagParams(null)
-    },
-    updateTagParams(_value) {
-      const updatedData = {
-        tagPlate: this.tagPlate,
-        tag1Group: this.tag1Group,
-        tag2Group: this.tag2Group,
-        walkingBy: this.walkingBy,
-        direction: this.direction,
-        offsetTagsBy: this.offsetTagsBy
-      }
-
-      this.$emit('tagparamsupdated', updatedData)
     }
   }
 }
